2일차 소스 작성중 커밋

+++ app/src/main/java/com/example/gurufinalproject/bean/MemberBean.java
@@ -3,6 +3,6 @@ package com.example.gurufinalproject.bean;
 public class MemberBean {
     public String name;
     public String userid;
-    public long userNum;
+    public int userNum;
     public String phoneNum;
 }
